requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
ffmpeg-python==0.2.0
filelock==3.16.1
fsspec==2024.12.0
future==1.0.0
idna==3.10
Jinja2==3.1.5
#llvmlite==0.43.0
llvmlite==0.42.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==2.8.8
#numba==0.60.0

numba==0.59.0


#numpy==2.0.2
#openai-whisper
pillow==11.0.0
regex==2024.11.6
requests==2.32.3
setuptools==75.6.0
sympy==1.13.1
tiktoken==0.8.0
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
#torch>=2.1
#torchaudio
#torch==2.1.0
#torchaudio==2.1.0
#TTS==0.24.2
#TTS
coqui-tts==0.24.2
scipy==1.11.2
numpy==1.25.2
soundfile
#librosa==0.10.1
librosa==0.10.1
unidic-lite
cutlet
opencv-python
scikit-image
basicsr==1.4.2
kornia==0.5.1
face-alignment==1.3.4
ninja==1.10.2.3
einops==0.6.0
facexlib==0.2.5
#dlib
gradio>=3.7.0
google-genai
